Store uids as int to support more sectors and units
Before, they were stored as short.  Wider uids use more space, but the
next commit will recover it by narrowing other members.

The use of short has always limited the number of ships, planes, land
units and nukes to SHRT_MAX (commonly 32768).  Only the most extreme
games ever came close.

Commit 49780e2c (v4.3.12) added struct sctstr member sct_uid to make
struct empobj member uid work for sectors.  This made the limit apply
to sectors as well.  We've had games with more than 32768 sectors.

+++ include/commodity.h
@@ -43,7 +43,7 @@
 struct comstr {
     /* initial part must match struct empobj */
     short ef_type;
-    short com_uid;
+    int com_uid;
     unsigned com_seqno;
     time_t com_timestamp;
     natid com_owner;
