+++ include/land.h
@@ -50,7 +50,7 @@
 struct lndstr {
     /* initial part must match struct empobj */
     short ef_type;
-    short lnd_uid;		/* unit id (land unit) */
+    int lnd_uid;		/* unit id (land #) */
     unsigned lnd_seqno;
     time_t lnd_timestamp;	/* Last time this unit was touched */
     natid lnd_own;		/* owner's country num */
@@ -66,7 +66,7 @@ struct lndstr {
     short lnd_mission;		/* mission code */
     short lnd_radius;		/* mission radius */
     /* end of part matching struct empobj */
-    short lnd_ship;		/* pointer to transporting ship */
+    int lnd_ship;		/* uid of transporting ship, or -1 */
     signed char lnd_harden;	/* fortification */
     short lnd_retreat;		/* retreat percentage */
     int lnd_rflags;		/* When do I retreat? */
@@ -75,7 +75,7 @@ struct lndstr {
     short lnd_item[I_MAX+1];	/* amount of items on board */
     short lnd_pstage;		/* plague stage */
     short lnd_ptime;		/* how many etus remain in this stage */
-    short lnd_land;		/* pointer to transporting unit */
+    int lnd_land;		/* uid of transporting land unit, or -1 */
     short lnd_access;		/* Last tick mob was updated (MOB_ACCESS) */
 };
