+++ include/loan.h
@@ -43,7 +43,7 @@
 struct lonstr {
     /* initial part must match struct empobj */
     short ef_type;
-    short l_uid;
+    int l_uid;
     unsigned l_seqno;
     time_t l_timestamp;
     /* end of part matching struct empobj */
