+++ include/sect.h
@@ -45,7 +45,7 @@
 struct sctstr {
     /* initial part must match struct empobj */
     short ef_type;
-    short sct_uid;		/* equals XYOFFSET(sct_x, sct_y) */
+    int sct_uid;		/* equals XYOFFSET(sct_x, sct_y) */
     unsigned sct_seqno;
     time_t sct_timestamp;	/* Last time this sector was written to */
     natid sct_own;		/* owner's country num */
